Work around a VirtualJaguar bug in font.c
VJ adjusts writes to 64-bit blitter registers such
that they are dword-swapped (E.g., a write to
B_PATD[0] goes to B_PATD[1] instead, and vice-
versa), but it fails to account for this when
reading back from these registers, at least when
not in phrase mode.

See blitter.cpp:BlitterWriteByte() and reads to
PATTERNDATA/B_PATD in blitter_generic(),
specifically the READ_RDATA() macro, which seems
to try to do the right thing in phrase mode, but
just passes the offset through in pixel mode.