Set phasers to stun! Now reading lightgun regs
Add a lightgun.s file that currently exports one
simple function: readgun.  It reads the light pen
registers and converts the raw values into a more
useful (x, y) pixels value.  Need to make another
pass on this by accepting the app's framebuffer
dimensions and PWIDTH as parameters, and use that
to convert the (x, y) scanline coordinates to
framebuffer coordinates, but this was a good place
for a checkpoint.