Convert lightpen registers to framebuffer (x,y)
Add code to convert the raw lightpen register
values all the way back to framebuffer
coordinates.  The code works on the assumption
that the framebuffer is centered on the active
video region, specifically centered using the
algorithm the standard Atari startup.s file
uses to set up the initial object processor
list.

This works well in general, but seems to require
a skew/calibration offset on the horizontal
values for some reason.  Perhaps this value just
accounts for propagation delay of the light-pen
signal, but its empirical value (~170 clocks)
seems rather high (works out to ~6.39us).

The math used here is as follows:

LPH = latched HC value, NOT horizontal position
      in pixels, as the manual states.

LPV = latched VC value, more or less matching
      the documentation.

Given LPH is is defined as a sampled value from
the HC register, we know it counts from zero to
the value in HP, or the horizontal period
register, which is the number of pixel/video
clock cycles in HALF of a scanline.  When it
wraps for the second half of a scanline, the
11th bit is set, so the first task we have is
to reconstruct the full-scanline pixel clock
count from LPH:

  raw = LPH
  hpcc = raw & ((1 << 10) - 1)
  if (raw & (1 << 1)) hpcc += HP

(hpcc: "Horizontal Pixel Clock Count")

However, we can't read HP (It's write-only).
Luckily, HP should be either NTSC_HMID or
PAL_HMID, so we can deduce it by checking the
NTSC/PAL config bit.

Next we want to convert from video clock ticks
to a pixel count, so we divide by the PWIDTH
field of VMODE.  The current code always assumes
PWIDTH4.  This should give us a scanline position
in pixels.

Now we need to map that pixel count onto the
framebuffer used by the object processor.  This
requires a lot of input variables, so I've
simplified as mentioned above by assuming a
framebuffer centered on the scanline horizontally.
The object processor's BITMAP object type contains
the field XPOS, which is defined as an offset in
pixels from the start of the line buffer.  OK, but
where is the line buffer positioned?  At HDB1/2 of
course!  Another write-only register, so back to
analyzing startup.s, we see it exports a_hdb, a
variable containing the value programmed in
HDB1 and HDB2.  Of course, a_hdb is in raw video
clocks, and XPOS is in pixels, so we need to take
care when combining them (Again, assuming PWIDTH4
here):

  framebuffer X origin = a_hdb + (XPOS * 4)

Where framebuffer X origin is in video clock
units.  How do we get XPOS?  We could pull it out
of the object list, pass it as a parameter, or
just re-calculate it from the framebuffer width
in pixels if we make the above assumptions about
a centered framebuffer, which is what I've done
here.

I call this value "xoff4" since it is the value
subtracted from the above hpcc value to obtain
the video clock offset within the framebuffer.
The result of that subtraction is then divided
by 4 to obtain the x coordinate within the
framebuffer image.

  fb_x = (hpcc - xoff4) / 4

Or that's how it *should* work in theory.  In
practice, I have to skew xoff4 a bit to the left:

  xoff4 -= 170
  ...
  fb_x = (hpcc - xoff4) / 4

...

The vertical position is much easier.  Just mask
off the "field" bit of LPV, divide it by two to
get the raw scanline count.

  vslc = (LPV & ((1 << 11) - 1)) / 2

(vslc: Vertical ScanLine Count)

From there, recalculate the YPOS object processor
value, which luckily is a raw offset from scanline
zero (though in half-scanlines, because everything
has to be hard on Jaguar) so the only inputs we
need here are the framebuffer height and the video
mode middle line, the latter of which is either
PAL_VMID or NTSC_VMID, which are also in half-
lines.

  framebuffer Y origin =
      (VMID - ((fb_height * 2) / 2)) / 2

or,

  framebuffer Y origin = (VMID - fb_height) / 2

We'll call this value yoff, similar to above for
the horizontal offset value, giving us the
following equation:

  fb_y = vslc - yoff

And there we have it, (fb_x, fb_y) is where the
light gun was pointed in the framebuffer! Simple,
right? Yeah... right...